requirements.txt
Flask==0.10.1
GoogleAppEngineMapReduce==1.9.22.0
google-api-python-client==1.5.3
